backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pypdf
python-docx
google-genai
python-dotenv

